datastore/fingerprinting/requirements.txt
Pillow
pytest
git+git://github.com/creativecommons/blockhash-python.git#egg=blockhash-python
